requirements.txt
web3==5.23.1
py-solc-x==1.1.1
python-dotenv==0.19.2